feat: GitHub 데이터 로직을 훅으로 분리하여 코드 구조 개선
