+++ assignment-3/src/pages/home/components/github.tsx
@@ -1,60 +1,24 @@
-import { useState } from "react";
-import { useGithubData } from "../hooks/use-github-data";
-import {
-  getRecentUsers,
-  removeRecentUser,
-  saveRecentUser,
-} from "../../../utils/local-storage";
+import { useGithubData } from "../hooks/get-github-data";
+import { useGithubLogic } from "../hooks/use-github-logic";
 import SearchBar from "../../../components/searchBar/search-bar";
 import * as styles from "./page.css";
 import Chip from "../../../components/chip/chip";
 import Card from "../../../components/card/card";
 
 const Github = () => {
-  const [recentUsers, setRecentUsers] = useState<string[]>([]);
-  const [user, setUser] = useState("");
-  const [queryUser, setQueryUser] = useState("");
-  const [isCardVisible, setIsCardVisible] = useState(false);
+  const {
+    user,
+    recentUsers,
+    queryUser,
+    isCardVisible,
+    handleOnChange,
+    handleKeyDown,
+    handleCloseCard,
+    handleDeleteChip,
+    handleChipClick,
+  } = useGithubLogic();
 
   const { data, isLoading, error } = useGithubData(queryUser);
-
-  const validateInput = (input: string) => input.trim().length > 0;
-
-  const updateQueryUser = (trimmedUser: string) => {
-    setQueryUser(trimmedUser);
-    setIsCardVisible(true);
-  };
-
-  const updateRecentUsers = (trimmedUser: string) => {
-    saveRecentUser(trimmedUser);
-    setRecentUsers(getRecentUsers());
-  };
-
-  const handleKeyDown = (e: React.KeyboardEvent<HTMLInputElement>) => {
-    if (e.key === "Enter" && validateInput(user)) {
-      const trimmedUser = user.trim();
-      updateQueryUser(trimmedUser);
-      updateRecentUsers(trimmedUser);
-    }
-  };
-
-  const handleOnChange = (e: React.ChangeEvent<HTMLInputElement>) => {
-    setUser(e.target.value);
-  };
-
-  const handleCloseCard = () => {
-    setIsCardVisible(false);
-    setUser("");
-  };
-
-  const handleDeleteChip = (keyword: string) => {
-    removeRecentUser(keyword);
-    setRecentUsers(getRecentUsers());
-  };
-
-  const handleChipClick = (keyword: string) => {
-    setQueryUser(keyword);
-  };
 
   return (
     <>
